Rename key iterators to indicate they are implementation details of Key.
Rename 'key.KeyIterator' -> 'key._KeyIterator'

Move / rename 'iterator.KeyDataIterator' -> 'key._KeyDataIterator'.

Fixes #244.
