+++ gcloud/storage/test_key.py
@@ -152,7 +152,7 @@ class Test_Key(unittest2.TestCase):
         bucket = _Bucket(connection)
         key = self._makeOne(bucket, KEY)
         fh = StringIO()
-        with _Monkey(MUT, KeyDataIterator=lambda self: iter(_CHUNKS)):
+        with _Monkey(MUT, _KeyDataIterator=lambda self: iter(_CHUNKS)):
             key.get_contents_to_file(fh)
         self.assertEqual(fh.getvalue(), ''.join(_CHUNKS))
 
@@ -165,7 +165,7 @@ class Test_Key(unittest2.TestCase):
         connection = _Connection()
         bucket = _Bucket(connection)
         key = self._makeOne(bucket, KEY)
-        with _Monkey(MUT, KeyDataIterator=lambda self: iter(_CHUNKS)):
+        with _Monkey(MUT, _KeyDataIterator=lambda self: iter(_CHUNKS)):
             with NamedTemporaryFile() as f:
                 key.get_contents_to_filename(f.name)
                 f.flush()
@@ -181,7 +181,7 @@ class Test_Key(unittest2.TestCase):
         connection = _Connection()
         bucket = _Bucket(connection)
         key = self._makeOne(bucket, KEY)
-        with _Monkey(MUT, KeyDataIterator=lambda self: iter(_CHUNKS)):
+        with _Monkey(MUT, _KeyDataIterator=lambda self: iter(_CHUNKS)):
             fetched = key.get_contents_as_string()
         self.assertEqual(fetched, ''.join(_CHUNKS))
 
@@ -591,11 +591,11 @@ class Test_Key(unittest2.TestCase):
         self.assertEqual(kw[0]['query_params'], {'projection': 'full'})
 
 
-class TestKeyIterator(unittest2.TestCase):
+class Test__KeyIterator(unittest2.TestCase):
 
     def _getTargetClass(self):
-        from gcloud.storage.key import KeyIterator
-        return KeyIterator
+        from gcloud.storage.key import _KeyIterator
+        return _KeyIterator
 
     def _makeOne(self, *args, **kw):
         return self._getTargetClass()(*args, **kw)
@@ -631,6 +631,172 @@ class TestKeyIterator(unittest2.TestCase):
         self.assertEqual(key.name, KEY)
 
 
+class Test__KeyDataIterator(unittest2.TestCase):
+
+    def _getTargetClass(self):
+        from gcloud.storage.key import _KeyDataIterator
+        return _KeyDataIterator
+
+    def _makeOne(self, *args, **kw):
+        return self._getTargetClass()(*args, **kw)
+
+    def test_ctor(self):
+        connection = _Connection()
+        key = _Key(connection)
+        iterator = self._makeOne(key)
+        self.assertTrue(iterator.key is key)
+        self.assertEqual(iterator._bytes_written, 0)
+        self.assertEqual(iterator._total_bytes, None)
+
+    def test__iter__(self):
+        response1 = _Response(status=200)
+        response1['content-range'] = '0-9/15'
+        response2 = _Response(status=200)
+        response2['content-range'] = '10-14/15'
+        connection = _Connection(
+            (response1, '0123456789'),
+            (response2, '01234'),
+        )
+        key = _Key(connection)
+        iterator = self._makeOne(key)
+        chunks = list(iterator)
+        self.assertEqual(len(chunks), 2)
+        self.assertEqual(chunks[0], '0123456789')
+        self.assertEqual(chunks[1], '01234')
+        self.assertEqual(iterator._bytes_written, 15)
+        self.assertEqual(iterator._total_bytes, 15)
+        kws = connection._requested
+        self.assertEqual(kws[0]['method'], 'GET')
+        self.assertEqual(kws[0]['url'],
+                         'http://example.com/b/name/o/key?alt=media')
+        self.assertEqual(kws[0]['headers'], {'Range': 'bytes=0-9'})
+        self.assertEqual(kws[1]['method'], 'GET')
+        self.assertEqual(kws[1]['url'],
+                         'http://example.com/b/name/o/key?alt=media')
+        self.assertEqual(kws[1]['headers'], {'Range': 'bytes=10-'})
+
+    def test_reset(self):
+        connection = _Connection()
+        key = _Key(connection)
+        iterator = self._makeOne(key)
+        iterator._bytes_written = 10
+        iterator._total_bytes = 1000
+        iterator.reset()
+        self.assertEqual(iterator._bytes_written, 0)
+        self.assertEqual(iterator._total_bytes, None)
+
+    def test_has_more_data_new(self):
+        connection = _Connection()
+        key = _Key(connection)
+        iterator = self._makeOne(key)
+        self.assertTrue(iterator.has_more_data())
+
+    def test_has_more_data_invalid(self):
+        connection = _Connection()
+        key = _Key(connection)
+        iterator = self._makeOne(key)
+        iterator._bytes_written = 10  # no _total_bytes.
+        self.assertRaises(ValueError, iterator.has_more_data)
+
+    def test_has_more_data_true(self):
+        connection = _Connection()
+        key = _Key(connection)
+        iterator = self._makeOne(key)
+        iterator._bytes_written = 10
+        iterator._total_bytes = 1000
+        self.assertTrue(iterator.has_more_data())
+
+    def test_has_more_data_false(self):
+        connection = _Connection()
+        key = _Key(connection)
+        iterator = self._makeOne(key)
+        iterator._bytes_written = 1000
+        iterator._total_bytes = 1000
+        self.assertFalse(iterator.has_more_data())
+
+    def test_get_headers_new(self):
+        connection = _Connection()
+        key = _Key(connection)
+        iterator = self._makeOne(key)
+        headers = iterator.get_headers()
+        self.assertEqual(len(headers), 1)
+        self.assertEqual(headers['Range'], 'bytes=0-9')
+
+    def test_get_headers_ok(self):
+        connection = _Connection()
+        key = _Key(connection)
+        iterator = self._makeOne(key)
+        iterator._bytes_written = 10
+        iterator._total_bytes = 1000
+        headers = iterator.get_headers()
+        self.assertEqual(len(headers), 1)
+        self.assertEqual(headers['Range'], 'bytes=10-19')
+
+    def test_get_headers_off_end(self):
+        connection = _Connection()
+        key = _Key(connection)
+        iterator = self._makeOne(key)
+        iterator._bytes_written = 95
+        iterator._total_bytes = 100
+        headers = iterator.get_headers()
+        self.assertEqual(len(headers), 1)
+        self.assertEqual(headers['Range'], 'bytes=95-')
+
+    def test_get_url(self):
+        connection = _Connection()
+        key = _Key(connection)
+        iterator = self._makeOne(key)
+        self.assertEqual(iterator.get_url(),
+                         'http://example.com/b/name/o/key?alt=media')
+
+    def test_get_next_chunk_underflow(self):
+        connection = _Connection()
+        key = _Key(connection)
+        iterator = self._makeOne(key)
+        iterator._bytes_written = iterator._total_bytes = 10
+        self.assertRaises(RuntimeError, iterator.get_next_chunk)
+
+    def test_get_next_chunk_200(self):
+        response = _Response(status=200)
+        response['content-range'] = '0-9/100'
+        connection = _Connection((response, 'CHUNK'))
+        key = _Key(connection)
+        iterator = self._makeOne(key)
+        chunk = iterator.get_next_chunk()
+        self.assertEqual(chunk, 'CHUNK')
+        self.assertEqual(iterator._bytes_written, len(chunk))
+        self.assertEqual(iterator._total_bytes, 100)
+        kw, = connection._requested
+        self.assertEqual(kw['method'], 'GET')
+        self.assertEqual(kw['url'],
+                         'http://example.com/b/name/o/key?alt=media')
+        self.assertEqual(kw['headers'], {'Range': 'bytes=0-9'})
+
+    def test_get_next_chunk_206(self):
+        response = _Response(status=206)
+        connection = _Connection((response, 'CHUNK'))
+        key = _Key(connection)
+        iterator = self._makeOne(key)
+        iterator._total_bytes = 1000
+        chunk = iterator.get_next_chunk()
+        self.assertEqual(chunk, 'CHUNK')
+        self.assertEqual(iterator._bytes_written, len(chunk))
+        kw, = connection._requested
+        self.assertEqual(kw['method'], 'GET')
+        self.assertEqual(kw['url'],
+                         'http://example.com/b/name/o/key?alt=media')
+        self.assertEqual(kw['headers'], {'Range': 'bytes=0-9'})
+
+    def test_get_next_chunk_416(self):
+        from gcloud.storage.exceptions import StorageError
+        response = _Response(status=416)
+        connection = _Connection((response, ''))
+        key = _Key(connection)
+        iterator = self._makeOne(key)
+        iterator._total_bytes = 1000
+        self.assertRaises(StorageError, iterator.get_next_chunk)
+
+
 class _Connection(object):
     API_BASE_URL = 'http://example.com'
 
@@ -664,6 +830,14 @@ class _Connection(object):
                 '&Expiration=%s' % expiration)
 
 
+class _Key(object):
+    CHUNK_SIZE = 10
+    path = '/b/name/o/key'
+
+    def __init__(self, connection):
+        self.connection = connection
+
+
 class _Bucket(object):
     path = '/b/name'
     name = 'name'
@@ -683,3 +857,9 @@ class _Bucket(object):
     def delete_key(self, key):
         del self._keys[key.name]
         self._deleted.append(key.name)
+
+
+class _Response(dict):
+    @property
+    def status(self):
+        return self['status']
